reference/eai-hub-agentic-starter-v2/tools/ngbse/Legacyfiles/NGBSE_15_0_Genesis/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ngbse"
version = "15.0.0"
description = "NGBSE 15.0 'Genesis' — Online OSINT engine with forecasting and governed automation"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name="NGBSE Team"}]
dependencies = [
    "requests>=2.32.3",
    "pyyaml>=6.0.2",
    "python-dateutil>=2.9.0",
    "stix2>=3.0.1",
    "python-docx>=1.1.2",
    "pydantic>=2.8.2",
    "beautifulsoup4>=4.12.3",
    "tldextract>=5.1.2",
]

[project.scripts]
ngbse = "ngbse.cli:main"
